Working product

+++ src/operations/softmax.h
@@ -4,6 +4,6 @@
 #include <math.h>
 #include <stdint.h>
 
-void softmax(const int32_t *data, int size, float *output);
+void softmax(const int8_t *data, int size, float *output);
 
 #endif // SOFTMAX_H
